refactor: renamed folder mapping-template to vtl
