+++ src/vtl/functionIsAdmin.ts
