--- test/main.test.ts
@@ -1,3 +1,0 @@
-test('Dummy', () => {
-  expect(true).toBeTruthy();
-});
